+++ src/main/java/org/sonar/plugins/delphi/symbol/resolve/ExpressionTypeResolver.java
@@ -230,6 +230,10 @@ public final class ExpressionTypeResolver {
         if (type == null) {
           type = findCurrentType(reference);
         }
+
+        if (type.isClassReference()) {
+          type = ((ClassReferenceType) type).classType();
+        }
       } else {
         type = handleNameOccurrence(occurrence);
       }