+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improve handling of cast expressions in `MemoryManagementRule`.
+- Improve primary expression type resolution around constructors.
 - Update various rule descriptions.
 
 ### Fixed